refactoring necessary files, moving them to new module and removing the pws modules
