+++ features/pws/.gitignore
@@ -1,0 +1,3 @@
+/build
+test/
+androidTest/
