+++ features/pws/src/main/AndroidManifest.xml
